requirements.txt
python-dotenv
requests
pynacl
